Simplified Chunk class. Removed unit tests for Chunk. Modified Table class and unit tests.
